refactor: split config types into config/types.rs
Move all struct/enum/macro definitions (Config, ProviderConfig, etc.)
from config/mod.rs into a dedicated config/types.rs. Keep mod.rs as a
thin facade with re-exports.
